refactor: reorganize packages and rename classes for naming consistency
